backend/requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
alembic==1.13.1
psycopg2-binary==2.9.9
pydantic[email]==2.7.0
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.0
python-multipart==0.0.9
emails==0.6
jinja2==3.1.3
python-dotenv==1.0.1
httpx==0.27.0
pytest==8.1.1
